htmlspecials.pixel() now no longer uses colored pixels, instead color is done via CSS. The URL for the one remaining transparent pixel can now be specified via src (either as an XML attribute or via the converter context).
Rename attrs methods with() and without() to withnames() and withoutnames()
(for Python 2.5 compatibility).

Use elinks instead of w3m for asText() and move/rename this method to a
function ll.xist.ns.html.astext().

Try to make XIST independent from PyXML (however PyXML is still required
for parsing via expat and for dtd2xsc.py (because this requires xmlproc)).

Remove the long dperecated method withSep().

Use Py_ssize_t in the C source where appropriate.

+++ src/ll/xist/helpers.h
@@ -8,11 +8,11 @@
 */
 
 
-static PyObject *STRINGLIB_NAME(PyObject *str, int inattr)
+static PyObject *STRINGLIB_NAME(PyObject *str, Py_ssize_t inattr)
 {
-	int i;
-	int oldsize;
-	int newsize = 0;
+	Py_ssize_t i;
+	Py_ssize_t oldsize;
+	Py_ssize_t newsize = 0;
 
 	oldsize = STRINGLIB_LEN(str);
 	for (i = 0; i < oldsize; ++i)
